make layouts for each route, move welcome page
