+++ app/layout.tsx
@@ -1,11 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "../styles/globals.css";
-import dynamic from "next/dynamic";
-
-interface LayoutProps {
-  includeSideMenu?: boolean;
-}
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -14,22 +9,16 @@ export const metadata: Metadata = {
   description: "Generated by create next app",
 };
 
-const SideMenu = dynamic(() => import("./_components/SideMenu"), {
-  ssr: false,
-});
-
 export default function RootLayout({
   children,
-  includeSideMenu = true,
 }: {
   children: React.ReactNode;
-  includeSideMenu?: boolean;
 }) {
   return (
-    <html>
+    <html lang="en">
       <body className="bg-color-default text-color-default">
-        {includeSideMenu && <SideMenu />}
-        {children}
+        {/* Layout UI */}
+        <main>{children}</main>
       </body>
     </html>
   );
