+++ app/welcome/_components/Hero.tsx
