+++ app/welcome/layout.tsx
@@ -1,0 +1,13 @@
+export default function WelcomeLayout({
+  children, // will be a page or nested layout
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <section>
+      {/* Include shared UI here e.g. a header or sidebar */}
+
+      {children}
+    </section>
+  );
+}
